Migrated SearchDemo to unified api

+++ SearchDemo/SearchDemo/SearchResultsController.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
-using System.Drawing;
+using Foundation;
+using UIKit;
+using CoreGraphics;
 
 namespace SearchDemo
 {
@@ -68,7 +68,7 @@ namespace SearchDemo
                 this.controller = controller;
             }
 
-            public override int RowsInSection (UITableView tableView, int section)
+            public override nint RowsInSection (UITableView tableView, nint section)
             {
                 return controller.searchResults.Count;
             }
